+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/baremetalmachine/_restart.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud baremetalmachine restart",
-    is_preview=True,
 )
 class Restart(AAZCommand):
     """Restart the provided bare metal machine.
@@ -23,9 +22,9 @@ class Restart(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-10-01-preview",
+        "version": "2025-02-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/baremetalmachines/{}/restart", "2024-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/baremetalmachines/{}/restart", "2025-02-01"],
         ]
     }
 
@@ -141,7 +140,7 @@ class Restart(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-10-01-preview",
+                    "api-version", "2025-02-01",
                     required=True,
                 ),
             }