+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/clustermanager/identity/_show.py
@@ -19,9 +19,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-10-01-preview",
+        "version": "2025-02-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clustermanagers/{}", "2024-10-01-preview", "identity"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clustermanagers/{}", "2025-02-01", "identity"],
         ]
     }
 
@@ -131,7 +131,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-10-01-preview",
+                    "api-version", "2025-02-01",
                     required=True,
                 ),
             }
@@ -175,6 +175,7 @@ class _ShowHelper:
     @classmethod
     def _build_schema_cluster_manager_read(cls, _schema):
         if cls._schema_cluster_manager_read is not None:
+            _schema.etag = cls._schema_cluster_manager_read.etag
             _schema.id = cls._schema_cluster_manager_read.id
             _schema.identity = cls._schema_cluster_manager_read.identity
             _schema.location = cls._schema_cluster_manager_read.location
@@ -188,6 +189,9 @@ class _ShowHelper:
         cls._schema_cluster_manager_read = _schema_cluster_manager_read = AAZObjectType()
 
         cluster_manager_read = _schema_cluster_manager_read
+        cluster_manager_read.etag = AAZStrType(
+            flags={"read_only": True},
+        )
         cluster_manager_read.id = AAZStrType(
             flags={"read_only": True},
         )
@@ -330,6 +334,7 @@ class _ShowHelper:
         tags = _schema_cluster_manager_read.tags
         tags.Element = AAZStrType()
 
+        _schema.etag = cls._schema_cluster_manager_read.etag
         _schema.id = cls._schema_cluster_manager_read.id
         _schema.identity = cls._schema_cluster_manager_read.identity
         _schema.location = cls._schema_cluster_manager_read.location