+++ src/datamigration/HISTORY.rst
@@ -4,6 +4,10 @@ Release History
 ===============
 
 =======
+1.0.0b4
+++++++
+* Updated the URL to download the SQL Server schema Zip to `https://aka.ms/sqlschemamigrationpackage`
+
 1.0.0b3
 ++++++
 * Updated the URL to download the SQL Assessment Zip to `https://aka.ms/sqlassessmentpackage`